+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/global-121/121-platform/compare/v1.116.2...main)
+## [Unreleased](https://github.com/global-121/121-platform/compare/v1.116.8...main)
 
 ### Removed
 
@@ -13,6 +13,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - AW-App interface code and all related references
 
 ---
+
+## [1.116.8](https://github.com/global-121/121-platform/compare/v1.116.7...v1.116.8)- 2024-04-19
+
+### Disabled
+
+- Quick search in frontend
+
+## [1.116.7](https://github.com/global-121/121-platform/compare/v1.116.6...v1.116.7)- 2024-04-18
+
+### Fixed
+
+- Performance issues for PV program
+- Heap out of memory when exporting
+- Quick search not applied when using 'Export filtered view'
 
 ## [1.116.6](https://github.com/global-121/121-platform/compare/v1.116.5...v1.116.6)- 2024-04-16